DataViews Extensibility: Allow unregistering the reorder-page action
